bot/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
discord.py==2.4.0
supabase==2.10.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
PyNaCl==1.5.0
